Updated AccessControlTestModule.js for Express 4.x
Updated AccessControlTestModule.js for Express 4.x.  This required some
changes to the way the code is written to support the changes in
express.js where express has become it's own web server middleware, and
the other components once part of express have been separated as
distinct middleware components.  These components need to be added to
the module and instantiated slightly different than in old versions of
express.